requirements.txt
numpy>=1.24
networkx>=3.0
pandas>=2.0
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.12
pulp>=2.7
pytest>=7.4
